Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,10 +3,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -28,8 +31,25 @@ public class UserService {
 		return repository.save(obj);
 	}
 
+	/* on this method, I needed use the "findById" to work exception ResourceNotFoundException.
+	 * the new versions of JPA, on deleteById method does findById implicited and not launch exception when
+	 * the object not exists on the database, returning silenty
+	 * the catch EmptyResultDataAccessException never will be launched in this context
+	 * */
 	public void delete(Long id) {
-		repository.deleteById(id);
+		User obj = findById(id);			
+		if (obj == null) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);			
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
